UC-343 move CKR_GENERAL_ERROR into the ‘re-establish session’ category in the error handler of the ubirch pkcs#11 crypto interface in order to prevent deadlock after restart of HSM module via csadm.
When the HSM module restarts, the session ID + key are deleted, while the TCP connection between client and module remain established. In this case, the old session ID + key become invalid, and an "invalid secure messaging ID"-error occurs when trying to create a signature. On client side, we only get a "CKR_GENERAL_ERROR" from the pkcs11 interface. In order to fix this, a new connection has to be established, i.e. tear down and re-establish the session on client side.

This commit fixes UC-343